move stuff from mixinextras to MixinHacks class, annotate internal stuff
