+++ build.gradle.kts
@@ -42,6 +42,7 @@ dependencies {
     implementation("org.spongepowered:mixin:0.8.5")
     implementation("org.ow2.asm:asm-tree:9.7")
     implementation("org.slf4j:slf4j-api:2.0.12")
+    implementation("org.jetbrains:annotations:24.1.0")
 
     "fabricImplementation"("net.fabricmc:fabric-loader:0.15.0")
 
@@ -54,6 +55,7 @@ dependencies {
 tasks.jar {
     from(fabric.output)
     from(forge.output)
+    from(neoforge.output)
 }
 
 tasks.register<Jar>("sourcesJar") {
