+++ src/fabric/java/com/moulberry/mixinconstraints/FabricAbstractionsImpl.java
@@ -5,6 +5,8 @@ import net.fabricmc.loader.api.FabricLoader;
 import net.fabricmc.loader.api.Version;
 import net.fabricmc.loader.api.VersionParsingException;
 
+import static com.moulberry.mixinconstraints.util.MixinHacks.unchecked;
+
 public class FabricAbstractionsImpl extends Abstractions {
 	@Override
 	protected boolean isDevEnvironment() {
@@ -34,7 +36,7 @@ public class FabricAbstractionsImpl extends Abstractions {
 			return (min == null || currentVersion.compareTo(min) >= 0) && (max == null || currentVersion.compareTo(max) <= 0);
 
 		} catch (VersionParsingException e) {
-			throw new RuntimeException(e);
+			throw unchecked(e);
 		}
 	}
 }
